chore: 기존 enter-reservation 테스트 페이지 및 컴포넌트, 타입선언 제거 (미사용) #124
