--- apps/client/src/app/(reservation)/enter-reservation-info/layout.tsx
@@ -1,9 +1,0 @@
-export default function layout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <div className="min-h-screen bg-inner-background-gray">{children}</div>
-  );
-}
