--- apps/client/src/components/pages/reservation/VehicleInfoBox.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function VehicleInfoBox() {
-  return <div>차량정보</div>;
-}
